Libnotify.C.Notify

Description

Low level bindings to libnotify

See also https://developer.gnome.org/libnotify/0.7/libnotify-notify.html. Haddocks here are mostly excerpts from there

Documentation

notify_init

Arguments

:: String
Application name. Should not be empty!

-> IO Bool

Initialize libnotify

This must be called before any other functions

notify_uninit :: IO ()

Uninitialize libnotify

notify_is_initted :: IO Bool

Get whether libnotify is initialized or not

notify_get_app_name :: IO String

Get the application name

Do not forget to call notify_init before calling this!

notify_set_app_name :: String -> IO ()

Set the application name

Do not forget to call notify_init before calling this!

notify_get_server_caps :: IO [String]

Return server capabilities

Synchronously queries the server for its capabilities

>>> notify_get_server_caps
["actions","body","body-markup","body-hyperlinks","icon-static","x-canonical-private-icon-only"]

data ServerInfo

Server information

Constructors

ServerInfo

Fields

name :: String

vendor :: String

version :: String

specVersion :: String

notify_get_server_info :: IO (Maybe ServerInfo)

Return server information

Synchronously queries the server for its information, specifically, the name, vendor, server version, and the version of the notifications specification that it is compliant with

>>> notify_get_server_info
Just (ServerInfo {name = "Xfce Notify Daemon", vendor = "Xfce", version = "0.2.4", specVersion = "0.9"})