System.Libnotify.Types

Description

Types used in other modules

Documentation

data Timeout

Timeout in seconds after which notification is closed. Although timeout does not work with notify-osd.

Constructors

Default
Default server timeout.

Custom Int
User defined timeout (in milliseconds).

Infinite
Notification will not expire until user pays attention to it.

data Urgency

Urgency can be used by the notification server to prioritize notifications. Although urgency does not work with notify-osd.

Constructors

Low
Low priority notification.

Normal
Default notification urgency.

Critical
Critical notification that requires immediate attention.

type Category = String

Category can be used by the notification server to filter or display the data in a certain way.

data ServerInfo

Server information.

Constructors

ServerInfo

Fields

name :: String

vendor :: String

version :: String

spec :: String